Fix potenial overflow in reverse4Bytes

+++ codeplug/misc.go
@@ -145,7 +145,7 @@ func intToBytes(i int, len int) []byte {
 // reverse4Bytes returns an integer containing the low four bytes of
 // the input in reverse order.
 func reverse4Bytes(in int) int {
-	uin := uint(in)
+	uin := uint64(in)
 	out := (uin & 0x000000ff) << 24
 	out |= (uin & 0x0000ff00) << 8
 	out |= (uin & 0x00ff0000) >> 8
